% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Easily create error messages for when arguments do not meet expectations}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{Argument, which will be a character string containing the name of the parameter.}

\item{must}{Character string which tells the requirement, such as 'be numeric'}

\item{not}{Argument variable, which will be used to tell if the parameter matches the requirement.}
}
\description{
Easily create error messages for when arguments do not meet expectations
}
