% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{add_direction_per_video_to_df}
\alias{add_direction_per_video_to_df}
\title{add_direction_per_video_to_df}
\usage{
add_direction_per_video_to_df(df)
}
\arguments{
\item{df}{df}
}
\value{
df_with_direction_metric
}
\description{
add_direction_per_video_to_df
}
