% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_survey_related_dir_list_to_env}
\alias{add_survey_related_dir_list_to_env}
\title{add_survey_related_dir_list_to_env}
\usage{
add_survey_related_dir_list_to_env(
  my_current_env,
  survey_version_name,
  survey_monkey_used,
  wave_names,
  should_create_nonexistant_dirs
)
}
\arguments{
\item{my_current_env}{my_current_env}

\item{survey_version_name}{survey_version_name}

\item{survey_monkey_used}{survey_monkey_used}

\item{wave_names}{wave_names}

\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}
}
\value{
invisible(my_current_env)
}
\description{
add_survey_related_dir_list_to_env
}
