% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_tables_to_analysis_wb}
\alias{add_tables_to_analysis_wb}
\title{add_tables_to_analysis_wb}
\usage{
add_tables_to_analysis_wb(
  wb,
  df,
  demographic_variable_names,
  where_the_tables_should_start_in_excel_row,
  where_the_tables_should_start_in_excel_col,
  sheet_name,
  group_by
)
}
\arguments{
\item{wb}{wb}

\item{df}{df}

\item{demographic_variable_names}{demographic_variable_names}

\item{where_the_tables_should_start_in_excel_row}{where_the_tables_should_start_in_excel_row}

\item{where_the_tables_should_start_in_excel_col}{where_the_tables_should_start_in_excel_col}

\item{sheet_name}{sheet_name}

\item{group_by}{group_by}
}
\value{
wb
}
\description{
add_tables_to_analysis_wb
}
