% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_traditional_project_dir_list_to_env}
\alias{add_traditional_project_dir_list_to_env}
\title{add_traditional_project_dir_list_to_env}
\usage{
add_traditional_project_dir_list_to_env(
  my_current_env,
  should_create_nonexistant_dirs
)
}
\arguments{
\item{my_current_env}{my_current_env}

\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}
}
\value{
invisible(my_current_env)
}
\description{
add_traditional_project_dir_list_to_env
}
