% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_order_for_other_labels}
\alias{adjust_order_for_other_labels}
\title{Adjust Order for Other Labels}
\usage{
adjust_order_for_other_labels(
  table,
  grouping_var_levels_included = TRUE,
  response_var_levels_included = TRUE
)
}
\arguments{
\item{table}{table}

\item{grouping_var_levels_included}{grouping_var_levels_included}

\item{response_var_levels_included}{response_var_levels_included}
}
\value{
table
}
\description{
Adjust Order for Other Labels
}
