% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_order_information_lookup_table_per_type}
\alias{adjust_order_information_lookup_table_per_type}
\title{adjust_order_information_lookup_table_per_type}
\usage{
adjust_order_information_lookup_table_per_type(
  table,
  order_information_lookup_table,
  type = "grouping_var_"
)
}
\arguments{
\item{table}{table}

\item{order_information_lookup_table}{order_information_lookup_table}

\item{type}{type}
}
\value{
table_2
}
\description{
adjust_order_information_lookup_table_per_type
}
