% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{calculate_direction_metric}
\alias{calculate_direction_metric}
\title{calculate_direction_metric}
\usage{
calculate_direction_metric(individual_immersion)
}
\arguments{
\item{individual_immersion}{individual_immersion}
}
\value{
direction
}
\description{
calculate_direction_metric
}
