% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capwcw_process_net_promoter}
\alias{capwcw_process_net_promoter}
\title{capwcw_process_net_promoter}
\usage{
capwcw_process_net_promoter(data, column_details)
}
\arguments{
\item{data}{data}

\item{column_details}{column_details}
}
\value{
data
}
\description{
capwcw_process_net_promoter
}
