% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_all_are_missing_from_row}
\alias{check_if_all_are_missing_from_row}
\title{check_if_all_are_missing_from_row}
\usage{
check_if_all_are_missing_from_row(df, list_of_variable_names)
}
\arguments{
\item{df}{df}

\item{list_of_variable_names}{list_of_variable_names}
}
\value{
is_all_column_data_missing_for_row
}
\description{
check_if_all_are_missing_from_row
}
