% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_and_process_with_column_workbook_list}
\alias{clean_and_process_with_column_workbook_list}
\title{Clean the survey data}
\usage{
clean_and_process_with_column_workbook_list(
  data,
  path_to_column_workbook,
  likert_indicator = NULL,
  name_of_column_details = "column_details"
)
}
\arguments{
\item{data}{data frame with the survey data}

\item{path_to_column_workbook}{Path to where the column details workbook is located.}

\item{likert_indicator}{character string that is used at the start of each likert variable name.}

\item{name_of_column_details}{column details df}
}
\value{
data_used
}
\description{
Clean the survey data
}
