% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_UPD_df}
\alias{clean_UPD_df}
\title{Clean the upd df}
\usage{
clean_UPD_df(
  user_data_df,
  age_group_levels = c("18-24", "25-34", "35-44", "45-54", "55-64", "65+")
)
}
\arguments{
\item{user_data_df}{dataframe containing the user participant data}

\item{age_group_levels}{what age groups should be included}
}
\value{
user_data
}
\description{
Clean the upd df
}
