% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{click_once_element_appears}
\alias{click_once_element_appears}
\title{click on element once it appears}
\usage{
click_once_element_appears(
  remDr,
  element,
  using = "css selector",
  sleep_time = 0.1,
  give_up_count_max = 999
)
}
\arguments{
\item{remDr}{Selenium client used to control the browser.}

\item{element}{A css element as a string.}

\item{using}{type of selector}

\item{sleep_time}{second to wait between loops.}

\item{give_up_count_max}{how many times to try the loops.}
}
\value{
ele
}
\description{
click on element once it appears
}
