% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_age_to_generation}
\alias{convert_age_to_generation}
\title{convert_age_to_generation}
\usage{
convert_age_to_generation(age_numeric, year_of_data_collection = NULL)
}
\arguments{
\item{age_numeric}{age_numeric}

\item{year_of_data_collection}{year_of_data_collection}
}
\value{
age_group
}
\description{
convert_age_to_generation
}
