% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_specific_column_names_to_data_type}
\alias{convert_specific_column_names_to_data_type}
\title{convert_specific_column_names_to_data_type}
\usage{
convert_specific_column_names_to_data_type(
  column_name_vector,
  text_names,
  numeric_names
)
}
\arguments{
\item{column_name_vector}{column_name_vector}

\item{text_names}{text_names}

\item{numeric_names}{numeric_names}
}
\value{
data_types
}
\description{
convert_specific_column_names_to_data_type
}
