% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_vars_to_factors_with_workbook}
\alias{convert_vars_to_factors_with_workbook}
\title{convert_vars_to_factors_with_workbook}
\usage{
convert_vars_to_factors_with_workbook(
  data,
  column_workbook_list,
  name_of_column_details
)
}
\arguments{
\item{data}{data}

\item{column_workbook_list}{column_workbook_list}

\item{name_of_column_details}{name_of_column_details}
}
\value{
df_to_use
}
\description{
convert_vars_to_factors_with_workbook
}
