% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_column_details_and_named_vectors_list}
\alias{create_column_details_and_named_vectors_list}
\title{Create the table that contains the information related to the variable names and column details}
\usage{
create_column_details_and_named_vectors_list(
  path_to_column_workbook,
  name_of_column_details = "column_details"
)

create_column_details_and_named_vectors_list(
  path_to_column_workbook,
  name_of_column_details = "column_details"
)
}
\arguments{
\item{path_to_column_workbook}{Path to where the column details workbook is located. This needs to be a character string and it needs to point to an existing file.}

\item{name_of_column_details}{Name of the column details sheet, which is defaulted to "column_details". This only needs to be changed if you do not want to use column details as the sheet name in the column details workbook.}
}
\value{
named_vectors_list

named_vectors_list
}
\description{
Create the table that contains the information related to the variable names and column details

Create the table that contains the information related to the variable names and column details
}
