% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_excel_file_paths_df}
\alias{create_excel_file_paths_df}
\title{create_excel_file_paths_df}
\usage{
create_excel_file_paths_df(
  data_folder_location,
  file_name_part,
  unique_name_part,
  file_ext,
  sheets_to_exclude
)
}
\arguments{
\item{data_folder_location}{data_folder_location}

\item{file_name_part}{file_name_part}

\item{unique_name_part}{unique_name_part}

\item{file_ext}{file_ext}

\item{sheets_to_exclude}{sheets_to_exclude}
}
\value{
file_paths_df_all_paths
}
\description{
create_excel_file_paths_df
}
