% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_file_paths_df}
\alias{create_file_paths_df}
\title{create_file_paths_df}
\usage{
create_file_paths_df(project_year, project_folder_name)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
file_paths_df
}
\description{
create_file_paths_df
}
