% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_frequency_table}
\alias{create_frequency_table}
\title{create_frequency_table}
\usage{
create_frequency_table(data, var1, var2_string)
}
\arguments{
\item{data}{data}

\item{var1}{var1}

\item{var2_string}{var2_string}
}
\value{
table
}
\description{
create_frequency_table
}
