% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_grouped_nps_stats_table}
\alias{create_grouped_nps_stats_table}
\title{create_grouped_nps_stats_table}
\usage{
create_grouped_nps_stats_table(df, grouping_var_name, response_var_name)
}
\arguments{
\item{df}{df}

\item{grouping_var_name}{grouping_var_name}

\item{response_var_name}{response_var_name}
}
\value{
table
}
\description{
create_grouped_nps_stats_table
}
