% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_grouped_percentages_table}
\alias{create_grouped_percentages_table}
\title{create_grouped_percentages_table}
\usage{
create_grouped_percentages_table(
  df,
  grouping_var_name,
  response_var_name,
  type = NULL,
  group_n_table = NULL
)
}
\arguments{
\item{df}{df}

\item{grouping_var_name}{grouping_var_name}

\item{response_var_name}{response_var_name}
}
\value{
grouped_percentage_table
}
\description{
create_grouped_percentages_table
}
