% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_immersion_animations}
\alias{create_immersion_animations}
\title{Create the immersion animation}
\usage{
create_immersion_animations(
  immersion_per_second_with_stats,
  colors_for_videos = NULL,
  raw_data_path,
  grouped_by
)
}
\arguments{
\item{immersion_per_second_with_stats}{Immersion per second data with the stimulus stats added}

\item{colors_for_videos}{What colors should be used}

\item{raw_data_path}{Where is the raw_data directory}

\item{grouped_by}{grouped_by}
}
\description{
Create the immersion animation
}
