% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{create_immersion_folder_structure_paths_list}
\alias{create_immersion_folder_structure_paths_list}
\title{create_immersion_folder_structure_paths_list}
\usage{
create_immersion_folder_structure_paths_list(project_year, project_folder_name)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
list_of_folders_paths
}
\description{
create_immersion_folder_structure_paths_list
}
