% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_immersion_video_vars_df}
\alias{create_immersion_video_vars_df}
\title{create_immersion_video_vars_df}
\usage{
create_immersion_video_vars_df(
  video_presentation_type_vec,
  video_name_vec,
  video_duration_vec
)
}
\arguments{
\item{video_presentation_type_vec}{video_presentation_type_vec}

\item{video_name_vec}{video_name_vec}

\item{video_duration_vec}{video_duration_vec}
}
\value{
immersion_video_vars_df
}
\description{
create_immersion_video_vars_df
}
