% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{create_initial_project_info}
\alias{create_initial_project_info}
\title{create_initial_project_info}
\usage{
create_initial_project_info(
  project_year,
  project_folder_name,
  study_type_selection,
  study_location_selection,
  n_participants_per_session_options,
  total_goal_sample_size
)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{study_type_selection}{study_type_selection}

\item{study_location_selection}{study_location_selection}

\item{n_participants_per_session_options}{n_participants_per_session_options}

\item{total_goal_sample_size}{total_goal_sample_size}
}
\value{
initial_project_info_df
}
\description{
create_initial_project_info
}
