% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_multiple_column_workbook_lists}
\alias{create_multiple_column_workbook_lists}
\title{create_multiple_column_workbook_lists}
\usage{
create_multiple_column_workbook_lists(my_env = my_current_env)
}
\arguments{
\item{my_env}{my_env}
}
\value{
column_workbook_lists
}
\description{
create_multiple_column_workbook_lists
}
