% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_nonexistant_files}
\alias{create_nonexistant_files}
\title{create_nonexistant_files}
\usage{
create_nonexistant_files(list_of_paths, list_of_file_types)
}
\arguments{
\item{list_of_paths}{list_of_paths}

\item{list_of_file_types}{list_of_file_types}
}
\description{
create_nonexistant_files
}
