% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_order_information_lookup_table}
\alias{create_order_information_lookup_table}
\title{create_order_information_lookup_table}
\usage{
create_order_information_lookup_table(
  column_workbook_list,
  name_of_column_details
)
}
\arguments{
\item{column_workbook_list}{column_workbook_list}

\item{name_of_column_details}{name_of_column_details}
}
\value{
full_lookup_table
}
\description{
create_order_information_lookup_table
}
