% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_path_to_lab_directory_in_dropbox}
\alias{create_path_to_lab_directory_in_dropbox}
\title{create_path_to_lab_directory_in_dropbox}
\usage{
create_path_to_lab_directory_in_dropbox(
  drop_box_name = "TCM Dropbox",
  dropbox_path_to_the_directory = "04 MDM Neuro-Fundraising Lab"
)
}
\arguments{
\item{drop_box_name}{Character string that contains the name of the dropbox directory you are using.}

\item{dropbox_path_to_the_directory}{Character string that contains the rest of the file path - once you are in dropbox}
}
\value{
created_path
}
\description{
create_path_to_lab_directory_in_dropbox
}
