% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_break_down_qualitative_combined}
\alias{create_power_bi_break_down_qualitative_combined}
\title{create_power_bi_break_down_qualitative_combined}
\usage{
create_power_bi_break_down_qualitative_combined(my_env = my_current_env)
}
\arguments{
\item{my_env}{my_env}
}
\value{
power_bi_break_down_qualitative_combined
}
\description{
create_power_bi_break_down_qualitative_combined
}
