% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_data}
\alias{create_power_bi_data}
\title{create_power_bi_data}
\usage{
create_power_bi_data(
  df,
  single_path_to_column_workbook,
  grouping_vars,
  response_var_type,
  qualitative_type = NULL
)
}
\arguments{
\item{df}{df}

\item{single_path_to_column_workbook}{single_path_to_column_workbook}

\item{grouping_vars}{grouping_vars}

\item{response_var_type}{response_var_type}

\item{qualitative_type}{qualitative_type}
}
\value{
combined_final_table
}
\description{
create_power_bi_data
}
