% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_data_nps_CALCULATED_TABLES}
\alias{create_power_bi_data_nps_CALCULATED_TABLES}
\title{create_power_bi_data_nps_CALCULATED_TABLES}
\usage{
create_power_bi_data_nps_CALCULATED_TABLES(
  df,
  column_workbook_list,
  grouping_vars,
  name_of_column_details
)
}
\arguments{
\item{df}{df}

\item{column_workbook_list}{column_workbook_list}

\item{grouping_vars}{grouping_vars}

\item{name_of_column_details}{name_of_column_details}
}
\value{
wave_tables_df
}
\description{
create_power_bi_data_nps_CALCULATED_TABLES
}
