% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_data_num}
\alias{create_power_bi_data_num}
\title{create_power_bi_data_num}
\usage{
create_power_bi_data_num(df, single_path_to_column_workbook, grouping_vars)
}
\arguments{
\item{df}{df}

\item{single_path_to_column_workbook}{single_path_to_column_workbook}

\item{grouping_vars}{grouping_vars}
}
\value{
table
}
\description{
create_power_bi_data_num
}
