% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_data_qualitative}
\alias{create_power_bi_data_qualitative}
\title{create_power_bi_data_qualitative}
\usage{
create_power_bi_data_qualitative(
  df,
  single_path_to_column_workbook,
  grouping_vars,
  qualitative_type = NULL,
  identifier = "respondent_id"
)
}
\arguments{
\item{df}{df}

\item{single_path_to_column_workbook}{single_path_to_column_workbook}

\item{grouping_vars}{grouping_vars}

\item{qualitative_type}{qualitative_type}

\item{identifier}{identifier}
}
\value{
combined_final_table
}
\description{
create_power_bi_data_qualitative
}
