% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_data_qualitative_CALCULATED_TABLES}
\alias{create_power_bi_data_qualitative_CALCULATED_TABLES}
\title{create_power_bi_data_qualitative_CALCULATED_TABLES}
\usage{
create_power_bi_data_qualitative_CALCULATED_TABLES(
  df = survey_data_for_power_bi_df,
  column_workbook_list,
  grouping_vars,
  name_of_column_details,
  path_to_qual_coding_data,
  qualitative_type = NULL,
  identifier
)
}
\arguments{
\item{df}{df}

\item{grouping_vars}{grouping_vars}

\item{path_to_qual_coding_data}{}

\item{qualitative_type}{qualitative_type}

\item{identifier}{identifier}
}
\value{
combined_final_table
}
\description{
create_power_bi_data_qualitative_CALCULATED_TABLES
}
