% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_multiple_choice}
\alias{create_power_bi_multiple_choice}
\title{create_power_bi_multiple_choice}
\usage{
create_power_bi_multiple_choice(my_env = my_current_env)
}
\arguments{
\item{my_env}{my_env}
}
\value{
power_bi_multiple_choice
}
\description{
create_power_bi_multiple_choice
}
