% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_overall_qualitative}
\alias{create_power_bi_overall_qualitative}
\title{create_power_bi_overall_qualitative}
\usage{
create_power_bi_overall_qualitative(my_env = my_current_env)
}
\arguments{
\item{my_env}{my_env}
}
\value{
invisible(my_env)
}
\description{
create_power_bi_overall_qualitative
}
