% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_power_bi_text_data}
\alias{create_power_bi_text_data}
\title{Reshape the text data so that it works with power bi}
\usage{
create_power_bi_text_data(text_survey_data, variables_to_include)
}
\arguments{
\item{text_survey_data}{survey data with only text related columns}

\item{variables_to_include}{Character vector with names of the variables to keep}
}
\value{
power_bi_text
}
\description{
Reshape the text data so that it works with power bi
}
