% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_specified_file_paths}
\alias{create_specified_file_paths}
\title{create_specified_file_paths}
\usage{
create_specified_file_paths(
  my_env = NULL,
  vector_of_path_names_or_objects_to_create = NULL,
  survey_directory_paths = NULL,
  column_names_paths = NULL,
  name_of_column_details = NULL
)
}
\arguments{
\item{my_env}{my_env}

\item{vector_of_path_names_or_objects_to_create}{vector_of_path_names_or_objects_to_create}

\item{survey_directory_paths}{survey_directory_paths}

\item{column_names_paths}{column_names_paths}

\item{name_of_column_details}{name_of_column_details}
}
\value{
invisible(my_env)
}
\description{
create_specified_file_paths
}
