% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_survey_folder_structure_paths_list_power_bi}
\alias{create_survey_folder_structure_paths_list_power_bi}
\title{create_survey_folder_structure_paths_list_power_bi}
\usage{
create_survey_folder_structure_paths_list_power_bi(
  project_year,
  project_folder_name,
  list_of_survey_version_names = list("All_Versions"),
  all_versions_in_same_analysis_folder = FALSE,
  storage_directory = "Dropbox (TCM)",
  should_create_survey_folders_if_they_do_not_exist = TRUE
)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{list_of_survey_version_names}{list_of_survey_version_names}

\item{all_versions_in_same_analysis_folder}{all_versions_in_same_analysis_folder}

\item{storage_directory}{storage_directory}

\item{should_create_survey_folders_if_they_do_not_exist}{should_create_survey_folders_if_they_do_not_exist}
}
\value{
survey_data_list
}
\description{
create_survey_folder_structure_paths_list_power_bi
}
