% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_survey_monkey_design_form_list}
\alias{create_survey_monkey_design_form_list}
\title{create_survey_monkey_design_form_list}
\usage{
create_survey_monkey_design_form_list(
  path,
  name_of_form = "Design Form",
  name_of_logic = "Logic",
  names_of_others_to_ignore = c("IGNORE")
)
}
\arguments{
\item{path}{path}

\item{name_of_form}{name_of_form}

\item{name_of_logic}{name_of_logic}

\item{names_of_others_to_ignore}{names_of_others_to_ignore}
}
\value{
named_vectors_list
}
\description{
create_survey_monkey_design_form_list
}
