% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_survey_related_dir_list}
\alias{create_survey_related_dir_list}
\title{create_survey_related_dir_list}
\usage{
create_survey_related_dir_list(
  survey_version_name = character(),
  survey_monkey_used = TRUE,
  wave_names = character(),
  should_create_nonexistant_dirs = TRUE
)
}
\arguments{
\item{survey_version_name}{survey_version_name}

\item{survey_monkey_used}{survey_monkey_used}

\item{wave_names}{wave_names}

\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}
}
\value{
invisible(list_of_paths)
}
\description{
create_survey_related_dir_list
}
