% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_table_single}
\alias{create_table_single}
\title{Create a single table}
\usage{
create_table_single(
  df,
  response_var_name,
  grouping_var_name,
  response_var_type
)
}
\arguments{
\item{df}{data frame with the survey data}

\item{response_var_name}{character with the response var name}

\item{grouping_var_name}{character with the grouping var name}

\item{response_var_type}{character string with the type of variable that is being processed}
}
\value{
table
}
\description{
Create a single table
}
