% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_table_single_mc_DEV}
\alias{create_table_single_mc_DEV}
\title{Create a single table Multiple Choice}
\usage{
create_table_single_mc_DEV(df, response_var_name, grouping_var_name)
}
\arguments{
\item{df}{data frame with the survey data}

\item{response_var_name}{character with the response var name}

\item{grouping_var_name}{character with the grouping var name}
}
\value{
table
}
\description{
Create a single table Multiple Choice
}
