% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_time_chr_string_for_file_names}
\alias{create_time_chr_string_for_file_names}
\title{create_time_chr_string_for_file_names}
\usage{
create_time_chr_string_for_file_names(format_string = "\%m\%d\%Y_\%H\%M")
}
\arguments{
\item{format_string}{format_string}
}
\value{
time_stamp
}
\description{
create_time_chr_string_for_file_names
}
