% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_traditional_project_dir_list}
\alias{create_traditional_project_dir_list}
\title{create_traditional_project_structure}
\usage{
create_traditional_project_dir_list(should_create_nonexistant_dirs = TRUE)
}
\arguments{
\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}
}
\value{
invisible(list_of_paths)
}
\description{
create_traditional_project_structure
}
