% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_user_participant_urls}
\alias{create_user_participant_urls}
\title{create_user_participant_urls}
\usage{
create_user_participant_urls(project_numbers)
}
\arguments{
\item{project_numbers}{Project numbers}
}
\value{
user_participant_urls
}
\description{
create_user_participant_urls
}
