% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_user_project_info_df}
\alias{create_user_project_info_df}
\title{create_user_project_info_df}
\usage{
create_user_project_info_df(user_project_name_vec, user_project_number_vec)
}
\arguments{
\item{user_project_name_vec}{user_project_name_vec}

\item{user_project_number_vec}{user_project_number_vec}
}
\value{
user_project_info_df
}
\description{
create_user_project_info_df
}
