% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_workbook_for_immersion_analysis}
\alias{create_workbook_for_immersion_analysis}
\title{create_workbook_for_immersion_analysis}
\usage{
create_workbook_for_immersion_analysis(df, path_to_immersion)
}
\arguments{
\item{df}{df}

\item{path_to_immersion}{path_to_immersion}
}
\value{
wb
}
\description{
create_workbook_for_immersion_analysis
}
