% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_schedule_data_from_user}
\alias{download_schedule_data_from_user}
\title{download_schedule_data_from_user}
\usage{
download_schedule_data_from_user(remDr, home_dir, project_numbers)
}
\arguments{
\item{remDr}{remDr}

\item{home_dir}{home_dir}

\item{project_numbers}{project_numbers}
}
\value{
file_paths
}
\description{
download_schedule_data_from_user
}
