% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_user_data_from_projects}
\alias{download_user_data_from_projects}
\title{Download the user data from a vector of project urls}
\usage{
download_user_data_from_projects(remDr, user_participant_urls)
}
\arguments{
\item{remDr}{engine}

\item{user_participant_urls}{user_participant_urls}
}
\description{
Download the user data from a vector of project urls
}
