% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_text_columns_correctly}
\alias{encode_text_columns_correctly}
\title{encode_text_columns_correctly}
\usage{
encode_text_columns_correctly(data)
}
\arguments{
\item{data}{df that has text columns}
}
\value{
data_used
}
\description{
encode_text_columns_correctly
}
