% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_column_details_question_type_df}
\alias{extract_column_details_question_type_df}
\title{Pull rows from the column details that are indicated as multiple choice.}
\usage{
extract_column_details_question_type_df(
  column_details_table,
  question_type_indicator = "mc"
)
}
\arguments{
\item{column_details_table}{column_details_table}

\item{question_type_indicator}{Character string that tells the program what data type is being evaluated and processed.}
}
\value{
column_details_question_type_df
}
\description{
Pull rows from the column details that are indicated as multiple choice.
}
