% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_vars_with_named_vectors}
\alias{factor_vars_with_named_vectors}
\title{factor_vars_with_named_vectors}
\usage{
factor_vars_with_named_vectors(
  df,
  list_with_variable_names_to_reorder,
  list_with_named_vectors
)
}
\arguments{
\item{df}{df}

\item{list_with_variable_names_to_reorder}{list_with_variable_names_to_reorder}

\item{list_with_named_vectors}{list_with_named_vectors}
}
\value{
df
}
\description{
factor_vars_with_named_vectors
}
