% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FILES_find_newest_file}
\alias{FILES_find_newest_file}
\title{FILES_find_newest_file
Function to take a path to a directory, and then it provides the most recent version of the file}
\usage{
FILES_find_newest_file(
  directory,
  file_type = ".csv",
  format_pattern = "_[0-9]{8}_[0-9]{4}",
  based_on_modiication_date = TRUE
)
}
\arguments{
\item{directory}{directory}

\item{file_type}{file_type}

\item{format_pattern}{format_pattern}

\item{based_on_modiication_date}{based_on_modiication_date}
}
\value{
newest_file
}
\description{
FILES_find_newest_file
Function to take a path to a directory, and then it provides the most recent version of the file
}
