% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{filter_out_anything_with_more_than_30_perc_loss}
\alias{filter_out_anything_with_more_than_30_perc_loss}
\title{filter_out_anything_with_more_than_30_perc_los}
\usage{
filter_out_anything_with_more_than_30_perc_loss(
  df,
  immersion_individual_summary
)
}
\arguments{
\item{df}{df}

\item{immersion_individual_summary}{immersion_individual_summary}
}
\value{
filtered_immersion_data
}
\description{
filter_out_anything_with_more_than_30_perc_los
}
