% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_video_times_with_list}
\alias{filter_video_times_with_list}
\title{filter_video_times_with_list}
\usage{
filter_video_times_with_list(df, list_of_video_names, list_of_video_seconds)
}
\arguments{
\item{df}{df}

\item{list_of_video_names}{list_of_video_names}

\item{list_of_video_seconds}{list_of_video_seconds}
}
\value{
filtered_immersion_data
}
\description{
filter_video_times_with_list
}
