% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_levels_and_order_in_table}
\alias{fix_levels_and_order_in_table}
\title{fix_levels_and_order_in_table}
\usage{
fix_levels_and_order_in_table(
  table,
  var_type,
  column_details,
  column_workbook_list,
  title_form = FALSE
)
}
\arguments{
\item{table}{table}

\item{var_type}{var_type}

\item{column_details}{column_details}

\item{column_workbook_list}{column_workbook_list}

\item{title_form}{title_form}
}
\value{
table
}
\description{
fix_levels_and_order_in_table
}
