% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FULL_global_coding}
\alias{FULL_global_coding}
\title{Fully read and process the global coding data.}
\usage{
FULL_global_coding(
  man_wd = NULL,
  file_part__sm_raw_folder = "Qualitative Coding/Version 2/global_variables/raw_data",
  Global_Coding_REFERENCE_file_path =
    "Qualitative Coding/Version 2/global_variables/Global_Coding_REFERENCE.xlsx",
  Global_Coding_CHANGE_LOG_file_path =
    "Qualitative Coding/Version 2/global_variables/Global_Coding_CHANGE_LOG.xlsx"
)
}
\arguments{
\item{man_wd}{man_wd}

\item{file_part__sm_raw_folder}{file_part__sm_raw_folder}

\item{Global_Coding_REFERENCE_file_path}{Global_Coding_REFERENCE_file_path}

\item{Global_Coding_CHANGE_LOG_file_path}{Global_Coding_CHANGE_LOG_file_path}
}
\value{
global_coding_to_save
}
\description{
Fully read and process the global coding data.
}
