% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content_Evaluation_Qualitative_Coding.R
\name{FULL_qualitative_coding_data}
\alias{FULL_qualitative_coding_data}
\title{FULL_qualitative_coding_data}
\usage{
FULL_qualitative_coding_data(
  data_folder_location,
  file_name_part,
  coder_names,
  ext__initial_name_qual_coding,
  text_names,
  numeric_names,
  multiple_choice_variables,
  category_variables,
  remove_empty_sheets = TRUE,
  other_vars_that_should_not_be_counted,
  sheets_to_exclude = c("Template", "Data Validation"),
  column_names_details_name = "Version 2_column_names.xlsx",
  create_log_doc = TRUE,
  storage_platform = "dropbox",
  storage_platform_name = "TCM Dropbox",
  group_dir_name = "04 MDM Neuro-Fundraising Lab",
  jobs_folder_name = "Research and Development/00 Jobs",
  project_year = 2024,
  project_folder_name = "003_RD_CodingVideoContent"
)
}
\arguments{
\item{data_folder_location}{data_folder_location}

\item{file_name_part}{file_name_part}

\item{coder_names}{coder_names}

\item{ext__initial_name_qual_coding}{ext__initial_name_qual_coding}

\item{text_names}{text_names}

\item{numeric_names}{numeric_names}

\item{multiple_choice_variables}{multiple_choice_variables}

\item{category_variables}{category_variables}

\item{remove_empty_sheets}{remove_empty_sheets}

\item{other_vars_that_should_not_be_counted}{other_vars_that_should_not_be_counted}

\item{sheets_to_exclude}{sheets_to_exclude}

\item{column_names_details_name}{column_names_details_name}

\item{create_log_doc}{create_log_doc}

\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}

\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
prepared_video_data
}
\description{
FULL_qualitative_coding_data
}
