% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_elements_from_raw_immersion_filename}
\alias{get_elements_from_raw_immersion_filename}
\title{Get the elements from the raw immersion filename}
\usage{
get_elements_from_raw_immersion_filename(raw_immersion_file_path)
}
\arguments{
\item{raw_immersion_file_path}{raw_immersion_file_path}
}
\value{
elements_from_raw_immersion_filename
}
\description{
Get the elements from the raw immersion filename
}
