% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_paths_and_column_data_from_excel_workbooks_list}
\alias{get_file_paths_and_column_data_from_excel_workbooks_list}
\title{get_file_paths_and_column_data_from_excel_workbooks_list}
\usage{
get_file_paths_and_column_data_from_excel_workbooks_list(
  video_coding_docs_file_paths,
  text_names,
  numeric_names,
  sheets_to_exclude
)
}
\arguments{
\item{video_coding_docs_file_paths}{video_coding_docs_file_paths}

\item{text_names}{text_names}

\item{numeric_names}{numeric_names}

\item{sheets_to_exclude}{sheets_to_exclude}
}
\value{
file_paths_df_all_paths
}
\description{
get_file_paths_and_column_data_from_excel_workbooks_list
}
