% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_paths_from_excel_workbooks_df_all_paths}
\alias{get_file_paths_from_excel_workbooks_df_all_paths}
\title{get_file_paths_from_excel_workbooks_df_all_paths}
\usage{
get_file_paths_from_excel_workbooks_df_all_paths(video_coding_docs_file_paths)
}
\arguments{
\item{video_coding_docs_file_paths}{video_coding_docs_file_paths}
}
\value{
file_paths_df_all_paths
}
\description{
get_file_paths_from_excel_workbooks_df_all_paths
}
