% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filtered_immersion_data}
\alias{get_filtered_immersion_data}
\title{Filter out excluded immersion data}
\usage{
get_filtered_immersion_data(all_immersion_data)
}
\arguments{
\item{all_immersion_data}{immersion data before exclusions}
}
\value{
filtered_immersion_data
}
\description{
Filter out excluded immersion data
}
