% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_column_details_from_dir}
\alias{get_full_column_details_from_dir}
\title{get_full_column_details_from_dir}
\usage{
get_full_column_details_from_dir(survey_directory_path)
}
\arguments{
\item{survey_directory_path}{survey_directory_path}
}
\value{
column_details
}
\description{
get_full_column_details_from_dir
}
