% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{get_immersion_individual_summary}
\alias{get_immersion_individual_summary}
\title{Obtain amount of missing data per rid, as well as mean and sd}
\usage{
get_immersion_individual_summary(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
immersion_individual_summary
}
\description{
Obtain amount of missing data per rid, as well as mean and sd
}
