% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{get_immersion_per_second}
\alias{get_immersion_per_second}
\title{Get Immersion Per Second Data}
\usage{
get_immersion_per_second(filtered_immersion_data, grouped_by)
}
\arguments{
\item{filtered_immersion_data}{data after filtering our the participants that need to be excluded.}

\item{grouped_by}{grouped_by}
}
\value{
immersion_per_second
}
\description{
Get Immersion Per Second Data
}
