% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_immersion_per_second_with_stats}
\alias{get_immersion_per_second_with_stats}
\title{Get Immersion per second with stats per stimulus}
\usage{
get_immersion_per_second_with_stats(filtered_immersion_data, grouped_by)
}
\arguments{
\item{filtered_immersion_data}{data after filtering our the participants that need to be excluded.}

\item{grouped_by}{grouped_by}
}
\value{
immersion_per_second_with_stats
}
\description{
Get Immersion per second with stats per stimulus
}
