% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_immersion_person_data}
\alias{get_immersion_person_data}
\title{Get Immersion Per Person Data}
\usage{
get_immersion_person_data(filtered_immersion_data, grouped_by)
}
\arguments{
\item{filtered_immersion_data}{data after filtering our the participants that need to be excluded.}

\item{grouped_by}{grouped_by}
}
\value{
immersion_person_data
}
\description{
Get Immersion Per Person Data
}
