% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{get_immersion_person_data_2}
\alias{get_immersion_person_data_2}
\title{get_immersion_person_data_2}
\usage{
get_immersion_person_data_2(filtered_immersion_data, grouped_by, extra_group)
}
\arguments{
\item{filtered_immersion_data}{filtered_immersion_data}

\item{grouped_by}{grouped_by}

\item{extra_group}{extra_group}
}
\value{
immersion_person_data
}
\description{
get_immersion_person_data_2
}
