% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{get_immersion_stimulis_stats}
\alias{get_immersion_stimulis_stats}
\title{Get Immersion stats for the stimuli}
\usage{
get_immersion_stimulis_stats(filtered_immersion_data, grouped_by)
}
\arguments{
\item{filtered_immersion_data}{data after filtering our the participants that need to be excluded.}

\item{grouped_by}{grouped_by}
}
\value{
immersion_stimulis_stats
}
\description{
Get Immersion stats for the stimuli
}
