% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_initial_read_data_from_excel_workbooks_with_converted_columns}
\alias{get_initial_read_data_from_excel_workbooks_with_converted_columns}
\title{get_initial_read_data_from_excel_workbooks_with_converted_columns}
\usage{
get_initial_read_data_from_excel_workbooks_with_converted_columns(
  path_to_excel_file,
  sheet_name
)
}
\arguments{
\item{path_to_excel_file}{path_to_excel_file}

\item{sheet_name}{sheet_name}
}
\value{
initial_data
}
\description{
get_initial_read_data_from_excel_workbooks_with_converted_columns
}
