% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_power_bi_parameters}
\alias{get_power_bi_parameters}
\title{get_power_bi_parameters}
\usage{
get_power_bi_parameters(
  project_year = 2023,
  project_folder_name = "230389_SDZ_BC__Quarterly Survey_OID1771",
  list_of_survey_version_names = list("Wave 1 - Member", "Wave 1 - Donor"),
  all_versions_in_same_analysis_folder = FALSE,
  other_vars_to_include,
  grouping_vars,
  spellcheck_replace_columns = TRUE,
  version_name_style,
  item_everyone_must_complete,
  remove_rids = FALSE,
  keep_rids = FALSE
)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{list_of_survey_version_names}{list_of_survey_version_names}

\item{all_versions_in_same_analysis_folder}{all_versions_in_same_analysis_folder}

\item{other_vars_to_include}{other_vars_to_include}

\item{grouping_vars}{grouping_vars}

\item{spellcheck_replace_columns}{spellcheck_replace_columns}
}
\description{
get_power_bi_parameters
}
