% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_session_information_from_list_of_file_names}
\alias{get_session_information_from_list_of_file_names}
\title{get_session_information_from_list_of_file_names}
\usage{
get_session_information_from_list_of_file_names(immersion_file_paths)
}
\arguments{
\item{immersion_file_paths}{immersion_file_paths}
}
\value{
session_information_from_file_names
}
\description{
get_session_information_from_list_of_file_names
}
