% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initiate_my_current_env}
\alias{initiate_my_current_env}
\title{initiate_my_current_env}
\usage{
initiate_my_current_env(
  storage_platform,
  storage_platform_name,
  group_dir_name,
  jobs_folder_name,
  project_year,
  project_folder_name
)
}
\arguments{
\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}

\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
invisible(my_current_env)
}
\description{
initiate_my_current_env
}
