% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initiate_project_and_design_details}
\alias{initiate_project_and_design_details}
\title{initiate_project_and_design_details}
\usage{
initiate_project_and_design_details()
}
\value{
user_credentials
}
\description{
initiate_project_and_design_details
}
