% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_text_present}
\alias{is_text_present}
\title{is_text_present}
\usage{
is_text_present(
  remDr,
  text_element,
  desired_text,
  num_of_iterations = 10,
  wait_time = 0.25
)
}
\arguments{
\item{remDr}{remDr}

\item{text_element}{text_element}

\item{desired_text}{desired_text}

\item{num_of_iterations}{num_of_iterations}

\item{wait_time}{wait_time}
}
\value{
found_text
}
\description{
is_text_present
}
