% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{make_the_immersion_animation}
\alias{make_the_immersion_animation}
\title{make_the_immersion_animation}
\usage{
make_the_immersion_animation(
  immersion_per_second_with_stats,
  the_label,
  video_color,
  min_y = 35,
  max_y = 75,
  min_x = 0,
  max_x = NULL,
  raw_data_path,
  grouped_by
)
}
\arguments{
\item{immersion_per_second_with_stats}{Average Immersion score for each second}

\item{the_label}{Label of the stimulus}

\item{video_color}{What colors should the images be}

\item{min_y}{min_y}

\item{max_y}{max_y}

\item{min_x}{min_x}

\item{max_x}{max_x}

\item{raw_data_path}{raw_data_path}

\item{grouped_by}{grouped_by}
}
\description{
make_the_immersion_animation
}
