% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_fill}
\alias{na_fill}
\title{Replace missing data with the appropriate label.}
\usage{
na_fill(x, firstBack = FALSE)
}
\arguments{
\item{x}{A vector of categorical values or labels.}

\item{firstBack}{A Boolean variable which should usually be false, so that it will keep the first NAs as NA. However, if I want to fill in the first NAs with the first label, I would need to make first back TRUE.}
}
\value{
x
}
\description{
Replaces NA values with the most recent valid label.
}
\author{
Brian Kissell
}
