% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{obtain_average_direction_per_video}
\alias{obtain_average_direction_per_video}
\title{obtain_average_direction_per_video}
\usage{
obtain_average_direction_per_video(df)
}
\arguments{
\item{df}{df}
}
\value{
average_video_direction_per_second
}
\description{
obtain_average_direction_per_video
}
