% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_column_details_table}
\alias{obtain_column_details_table}
\title{Extract the table that contains the column_details}
\usage{
obtain_column_details_table(
  column_workbook_list,
  name_of_column_details = "column_details"
)
}
\arguments{
\item{column_workbook_list}{List created by `create_column_details_and_named_vectors_list`, which has the column names along with all named vectors. This list should only contain the data for 1 column details file.}

\item{name_of_column_details}{Name of the column details sheet, which is defaulted to "column_details". This only needs to be changed if you do not want to use column details as the sheet name in the column details workbook.}
}
\value{
column_details_table
}
\description{
Extract the table that contains the column_details
}
