% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_column_name_path}
\alias{obtain_column_name_path}
\title{Obtain column name paths for all survey versions}
\usage{
obtain_column_name_path(survey_directory_path)
}
\arguments{
\item{survey_directory_path}{survey_directory_path}
}
\value{
column_name_path
}
\description{
Obtain column name paths for all survey versions
}
