% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_connection_to_zip_files_for_versions}
\alias{obtain_connection_to_zip_files_for_versions}
\title{Obtain connection to zip files for versions}
\usage{
obtain_connection_to_zip_files_for_versions(newest_files_for_project)
}
\arguments{
\item{newest_files_for_project}{newest_files_for_project}
}
\value{
connection_to_zip_files
}
\description{
Obtain connection to zip files for versions
}
