% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_current_file_path_parameters_df}
\alias{obtain_current_file_path_parameters_df}
\title{obtain_current_file_path_parameters_df}
\usage{
obtain_current_file_path_parameters_df(file_paths_df)
}
\arguments{
\item{file_paths_df}{file_paths_df}
}
\value{
current_file_path_parameters_df
}
\description{
obtain_current_file_path_parameters_df
}
