% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_immersion_video_details_df}
\alias{obtain_immersion_video_details_df}
\title{obtain_immersion_video_details_df}
\usage{
obtain_immersion_video_details_df(
  immersion_video_vars_df,
  total_goal_sample_size
)
}
\arguments{
\item{immersion_video_vars_df}{immersion_video_vars_df}

\item{total_goal_sample_size}{total_goal_sample_size}
}
\value{
immersion_video_details_df
}
\description{
obtain_immersion_video_details_df
}
