% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{obtain_individual_immersion_summary_with_rid_video_video_condition_order}
\alias{obtain_individual_immersion_summary_with_rid_video_video_condition_order}
\title{obtain_individual_immersion_summary_with_rid_video_video_condition_order}
\usage{
obtain_individual_immersion_summary_with_rid_video_video_condition_order(data)
}
\arguments{
\item{data}{data}
}
\value{
immersion_individual_summary
}
\description{
obtain_individual_immersion_summary_with_rid_video_video_condition_order
}
