% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_initial_column_names_for_version}
\alias{obtain_initial_column_names_for_version}
\title{obtain_initial_column_names_for_version}
\usage{
obtain_initial_column_names_for_version(
  version_column_names_paths,
  the_column_names_file_exists
)
}
\arguments{
\item{version_column_names_paths}{version_column_names_paths}

\item{the_column_names_file_exists}{the_column_names_file_exists}
}
\value{
initial_column_names_for_version
}
\description{
obtain_initial_column_names_for_version
}
