% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{obtain_initial_project_info_parameters_df}
\alias{obtain_initial_project_info_parameters_df}
\title{obtain_initial_project_info_parameters_df}
\usage{
obtain_initial_project_info_parameters_df(initial_project_info_df)
}
\arguments{
\item{initial_project_info_df}{initial_project_info_df}
}
\value{
initial_project_info_parameters_df
}
\description{
obtain_initial_project_info_parameters_df
}
