% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_location_from_table_list}
\alias{obtain_location_from_table_list}
\title{obtain_location_from_table_list}
\usage{
obtain_location_from_table_list(
  table_list,
  initial_value = 2,
  spacing_size = 1,
  direction = "horizontal"
)
}
\arguments{
\item{table_list}{table_list}

\item{initial_value}{initial_value}

\item{spacing_size}{spacing_size}

\item{direction}{direction}
}
\value{
location_values_for_list_of_tables
}
\description{
obtain_location_from_table_list
}
