% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_path_for_user_credentials}
\alias{obtain_path_for_user_credentials}
\title{obtain_path_for_user_credentials}
\usage{
obtain_path_for_user_credentials(path_part_to_adjust)
}
\arguments{
\item{path_part_to_adjust}{path_part_to_adjust}
}
\value{
path_to_credentials
}
\description{
obtain_path_for_user_credentials
}
