% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_project_design_information_xlsx_path}
\alias{obtain_project_design_information_xlsx_path}
\title{obtain_project_design_information_xlsx_path}
\usage{
obtain_project_design_information_xlsx_path(project_year, project_folder_name)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
project_design_information_xlsx_path
}
\description{
obtain_project_design_information_xlsx_path
}
