% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_the_connection_to_zip_file_collector_list}
\alias{obtain_the_connection_to_zip_file_collector_list}
\title{obtain_the_connection_to_zip_file_collector_list}
\usage{
obtain_the_connection_to_zip_file_collector_list(newest_files_for_project)
}
\arguments{
\item{newest_files_for_project}{newest_files_for_project}
}
\value{
connection_to_zip_files
}
\description{
obtain_the_connection_to_zip_file_collector_list
}
