% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_environmental_variables_for_current_project}
\alias{prepare_environmental_variables_for_current_project}
\title{prepare_environmental_variables_for_current_project}
\usage{
prepare_environmental_variables_for_current_project(
  environment_name,
  storage_platform,
  storage_platform_name,
  group_dir_name,
  jobs_folder_name,
  project_year,
  project_folder_name,
  should_create_nonexistant_dirs,
  project_numbers
)
}
\arguments{
\item{environment_name}{environment_name}

\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}

\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}

\item{project_numbers}{project_numbers}
}
\value{
environment_name
}
\description{
prepare_environmental_variables_for_current_project
}
