% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_UPD_file_paths_for_processing}
\alias{prepare_UPD_file_paths_for_processing}
\title{prepare_UPD_file_paths_for_processing
Prepare User Participant Data File Paths for Processing}
\usage{
prepare_UPD_file_paths_for_processing(environment_name)
}
\arguments{
\item{environment_name}{environment_name}
}
\value{
invisible(environment_name)
}
\description{
prepare_UPD_file_paths_for_processing
Prepare User Participant Data File Paths for Processing
}
