% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_VAR_ORDER_variables_single}
\alias{prepare_VAR_ORDER_variables_single}
\title{Prepare the grouping and response variables}
\usage{
prepare_VAR_ORDER_variables_single(df, response_var_name, grouping_var_name)
}
\arguments{
\item{df}{data frame with the survey data}

\item{response_var_name}{character with the response var name}

\item{grouping_var_name}{character with the grouping var name}
}
\value{
df
}
\description{
Prepare the grouping and response variables
}
