% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_version_directory_paths}
\alias{prepare_version_directory_paths}
\title{Prepare survey version directory paths}
\usage{
prepare_version_directory_paths(survey_directory_path)
}
\arguments{
\item{survey_directory_path}{path to the surveymonkeydata folder}
}
\value{
version_directories
}
\description{
Prepare survey version directory paths
}
