% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content_Evaluation_Qualitative_Coding.R
\name{prepare_video_duration_data}
\alias{prepare_video_duration_data}
\title{prepare_video_duration_data}
\usage{
prepare_video_duration_data(data_for_calcs_wide, all_vars_to_count_duration)
}
\arguments{
\item{data_for_calcs_wide}{data_for_calcs_wide}

\item{all_vars_to_count_duration}{all_vars_to_count_duration}
}
\value{
prepared_video_duration_data
}
\description{
prepare_video_duration_data
}
