% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_a_select_all_response_as_ones_and_zeros}
\alias{process_a_select_all_response_as_ones_and_zeros}
\title{process_a_select_all_response_as_ones_and_zeros}
\usage{
process_a_select_all_response_as_ones_and_zeros(
  is_all_column_data_missing_for_row,
  variable
)
}
\arguments{
\item{is_all_column_data_missing_for_row}{is_all_column_data_missing_for_row}

\item{variable}{variable}
}
\value{
processed_var
}
\description{
process_a_select_all_response_as_ones_and_zeros
}
