% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_and_clean_power_bi_survey_data}
\alias{process_and_clean_power_bi_survey_data}
\title{process_and_clean_power_bi_survey_data}
\usage{
process_and_clean_power_bi_survey_data(
  survey_directory_path,
  spellcheck_replace_columns = TRUE,
  spellcheck_column_paths = NULL,
  all_versions_in_same_analysis_folder = TRUE,
  path_to_column_workbook = NULL,
  version_name_style = "wave - condition",
  likert_indicator = "ec__",
  item_everyone_must_complete = "annual_household_income",
  remove_rids = FALSE,
  keep_rids = FALSE
)
}
\arguments{
\item{survey_directory_path}{survey_directory_path}

\item{spellcheck_replace_columns}{spellcheck_replace_columns}

\item{spellcheck_column_paths}{spellcheck_column_paths}

\item{all_versions_in_same_analysis_folder}{all_versions_in_same_analysis_folder}

\item{path_to_column_workbook}{path_to_column_workbook}

\item{version_name_style}{version_name_style}

\item{likert_indicator}{likert_indicator}

\item{item_everyone_must_complete}{item_everyone_must_complete}

\item{remove_rids}{remove_rids}

\item{keep_rids}{keep_rids}
}
\value{
survey_data_list
}
\description{
process_and_clean_power_bi_survey_data
}
