% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_numeric_vars}
\alias{process_numeric_vars}
\title{Process the data for all of the numeric variables}
\usage{
process_numeric_vars(
  data_used,
  column_details,
  likert_indicator = likert_indicator
)
}
\arguments{
\item{data_used}{data frame with the survey data}

\item{column_details}{column details df}

\item{likert_indicator}{character string that is used at the start of each likert variable name.}
}
\value{
data_used
}
\description{
Process the data for all of the numeric variables
}
