% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_scraped_schedules}
\alias{process_scraped_schedules}
\title{Process scraped schedules}
\usage{
process_scraped_schedules(home_dir)
}
\arguments{
\item{home_dir}{home_dir}
}
\value{
scrape_df
}
\description{
Process scraped schedules
}
