% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_select_all_vars}
\alias{process_select_all_vars}
\title{Process the data for all of the select all questions}
\usage{
process_select_all_vars(
  df,
  column_details_question_type_df,
  column_workbook_list
)
}
\arguments{
\item{df}{data frame that contains the survey data}

\item{column_details_question_type_df}{data frame that contains the column details for a specified question type}

\item{column_workbook_list}{List with all of the named vectors}
}
\value{
df
}
\description{
Process the data for all of the select all questions
}
