% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_select_all_vars_power_bi}
\alias{process_select_all_vars_power_bi}
\title{process_select_all_vars_power_bi}
\usage{
process_select_all_vars_power_bi(
  data,
  column_workbook_list,
  name_of_column_details
)
}
\arguments{
\item{data}{data}

\item{column_workbook_list}{column_workbook_list}

\item{name_of_column_details}{name_of_column_details}
}
\value{
df_to_use
}
\description{
process_select_all_vars_power_bi
}
