% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvd__further_cleaning}
\alias{pvd__further_cleaning}
\title{pvd__further_cleaning}
\usage{
pvd__further_cleaning(
  ind_sheet_coding_data_df,
  multiple_choice_variables,
  category_variables
)
}
\arguments{
\item{ind_sheet_coding_data_df}{ind_sheet_coding_data_df}

\item{multiple_choice_variables}{multiple_choice_variables}

\item{category_variables}{category_variables}
}
\value{
sheet_coding_data_df
}
\description{
pvd__further_cleaning
}
