% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content_Evaluation_Qualitative_Coding.R
\name{pvd__read_and_process_sheet_data}
\alias{pvd__read_and_process_sheet_data}
\title{pvd__read_and_process_sheet_data}
\usage{
pvd__read_and_process_sheet_data(
  individual_file_information_df,
  column_workbook_list
)
}
\arguments{
\item{individual_file_information_df}{individual_file_information_df}

\item{column_workbook_list}{column_workbook_list}
}
\value{
sheet_coding_data
}
\description{
pvd__read_and_process_sheet_data
}
