% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualitative_coding_system}
\alias{qualitative_coding_system}
\title{qualitative_coding_system}
\usage{
qualitative_coding_system(
  qual_path_list = my_current_env$qualitative_coding_data_path_list,
  processed_survey_data = my_current_env$survey_data_for_power_bi_df,
  column_workbook_lists_single = my_current_env$column_workbook_lists_single,
  grouping_vars,
  name_of_column_details,
  path_to_qual_coding_data_list = my_current_env$qualitative_coding_data_path_list,
  write_data,
  power_bi_overall_qualitative_path = my_current_env$power_bi_overall_qualitative_path,
  power_bi_break_down_qualitative_path =
    my_current_env$power_bi_break_down_qualitative_path
)
}
\arguments{
\item{qual_path_list}{qual_path_list}

\item{processed_survey_data}{processed_survey_data}

\item{column_workbook_lists_single}{column_workbook_lists_single}

\item{grouping_vars}{grouping_vars}

\item{name_of_column_details}{name_of_column_details}

\item{path_to_qual_coding_data_list}{path_to_qual_coding_data_list}

\item{write_data}{write_data}

\item{power_bi_overall_qualitative_path}{power_bi_overall_qualitative_path}

\item{power_bi_break_down_qualitative_path}{power_bi_break_down_qualitative_path}
}
\value{
power_bi_overall_qualitative_combined
}
\description{
qualitative_coding_system
}
