% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_and_process_UPD}
\alias{read_and_process_UPD}
\title{read_and_process_UPD
Read and process user data}
\usage{
read_and_process_UPD(upd_paths, Clip_data = FALSE)
}
\arguments{
\item{upd_paths}{paths for the data to read}

\item{Clip_data}{should it put the files in the clipboard}
}
\value{
user_participant_data
}
\description{
read_and_process_UPD
Read and process user data
}
