% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_in_zip}
\alias{read_csv_in_zip}
\title{read_csv_in_zip}
\usage{
read_csv_in_zip(connection_to_zip_files, initial_column_names_for_version)
}
\arguments{
\item{connection_to_zip_files}{connection_to_zip_files}

\item{initial_column_names_for_version}{initial_column_names_for_version}
}
\value{
survey_data
}
\description{
read_csv_in_zip
}
