% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_in_zip_single}
\alias{read_csv_in_zip_single}
\title{read_csv_in_zip_single}
\usage{
read_csv_in_zip_single(
  connection_to_zip_file,
  initial_column_names_for_version
)
}
\arguments{
\item{initial_column_names_for_version}{initial_column_names_for_version}

\item{connection_to_zip_files}{connection_to_zip_files}
}
\value{
survey_data
}
\description{
read_csv_in_zip_single
}
