% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_immersion_files}
\alias{read_immersion_files}
\title{Read and prepare the immersion data}
\usage{
read_immersion_files(
  raw_data_path,
  allow_n_repeats = 6,
  save_file_name_session_information = FALSE,
  remove_end_and_init_labels = TRUE,
  list_of_bs_video_names = NULL,
  list_of_bs_video_seconds = NULL,
  list_of_ws_video_names = NULL,
  list_of_ws_video_seconds = NULL,
  filter_video_lengths = TRUE
)
}
\arguments{
\item{raw_data_path}{This is where the raw data files are located}

\item{allow_n_repeats}{allow_n_repeats}

\item{save_file_name_session_information}{save_file_name_session_information}

\item{remove_end_and_init_labels}{remove_end_and_init_labels}

\item{list_of_bs_video_names}{list_of_bs_video_names}

\item{list_of_bs_video_seconds}{list_of_bs_video_seconds}

\item{list_of_ws_video_names}{list_of_ws_video_names}

\item{list_of_ws_video_seconds}{list_of_ws_video_seconds}

\item{filter_video_lengths}{filter_video_lengths}
}
\value{
immersion_data
}
\description{
Read and prepare the immersion data
}
