% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_immersion_options}
\alias{read_immersion_options}
\title{read_immersion_options}
\usage{
read_immersion_options(
  give_path = NULL,
  filter_out_bad = TRUE,
  show_inclusion_summary = TRUE
)
}
\arguments{
\item{give_path}{path, explorer if null}

\item{filter_out_bad}{filter data?}

\item{show_inclusion_summary}{show_inclusion_summary}
}
\value{
immersion_data
}
\description{
read_immersion_options
}
