% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_survey_data}
\alias{read_survey_data}
\title{read_survey_data}
\usage{
read_survey_data(survey_monkey_raw_data_path)
}
\arguments{
\item{survey_monkey_raw_data_path}{}
}
\value{
survey_data_list
}
\description{
read_survey_data
}
