% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_survey_data_for_power_bi}
\alias{read_survey_data_for_power_bi}
\title{read_survey_data_for_power_bi}
\usage{
read_survey_data_for_power_bi(survey_directory_path)
}
\arguments{
\item{survey_directory_path}{survey_directory_path}
}
\value{
survey_data_list
}
\description{
read_survey_data_for_power_bi
}
