% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_survey_data_for_power_bi_single}
\alias{read_survey_data_for_power_bi_single}
\title{read_survey_data_for_power_bi_single}
\usage{
read_survey_data_for_power_bi_single(
  survey_directory_path,
  file_ext = ".zip",
  format_pattern = "_[0-9]{8}_[0-9]{4}",
  survey_version_name,
  column_workbook_list
)
}
\arguments{
\item{survey_directory_path}{survey_directory_path}
}
\value{
survey_data_list
}
\description{
read_survey_data_for_power_bi_single
}
