% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_survey_monkey_data}
\alias{read_survey_monkey_data}
\title{read_survey_monkey_data}
\usage{
read_survey_monkey_data(
  man_wd = NULL,
  file_part__sm_raw_folder = "Qualitative Coding/Version 2/global_variables/raw_data"
)
}
\arguments{
\item{man_wd}{man_wd}

\item{file_part__sm_raw_folder}{file_part__sm_raw_folder}
}
\value{
survey_data_list
}
\description{
read_survey_monkey_data
}
