% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_as_factor}
\alias{recode_as_factor}
\title{recode_as_factor}
\usage{
recode_as_factor(variable, named_vec_levels)
}
\arguments{
\item{variable}{variable}

\item{named_vec_levels}{named_vec_levels}
}
\value{
variable
}
\description{
recode_as_factor
}
