% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_immersion.R
\name{replace_repeats_over_n_with_na}
\alias{replace_repeats_over_n_with_na}
\title{Create a function to replaces any values that are repeated more than n times with an NA}
\usage{
replace_repeats_over_n_with_na(var, n = 6)
}
\arguments{
\item{var}{Provide the name of the variable as a string}

\item{n}{Provide the number of repeated values to keep}
}
\value{
cleaned_values
}
\description{
Create a function to replaces any values that are repeated more than n times with an NA
}
