% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_spelling_with_excel_workbook}
\alias{replace_spelling_with_excel_workbook}
\title{replace_spelling_with_excel_workbook}
\usage{
replace_spelling_with_excel_workbook(df, spellchecked_path)
}
\arguments{
\item{df}{df}

\item{spellchecked_path}{spellchecked_path}
}
\value{
df_to_fix
}
\description{
replace_spelling_with_excel_workbook
}
