% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RUN_AND_CLIP_USER_PARTICIPANT_DATA}
\alias{RUN_AND_CLIP_USER_PARTICIPANT_DATA}
\title{RUN_AND_CLIP_USER_PARTICIPANT_DATA}
\usage{
RUN_AND_CLIP_USER_PARTICIPANT_DATA(
  environment_name,
  remDr,
  user_participant_urls
)
}
\arguments{
\item{environment_name}{environment_name}

\item{remDr}{remDr}

\item{user_participant_urls}{user_participant_urls}
}
\value{
invisible(environment_name)
}
\description{
RUN_AND_CLIP_USER_PARTICIPANT_DATA
}
