% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_create_table_with_var_names_and_orders}
\alias{run_create_table_with_var_names_and_orders}
\title{run_create_table_with_var_names_and_orders}
\usage{
run_create_table_with_var_names_and_orders(
  response_var_type,
  df,
  response_var_names,
  response_var_used_order,
  grouping_vars,
  grouping_var_used_order
)
}
\arguments{
\item{response_var_type}{response_var_type}

\item{df}{df}

\item{response_var_names}{response_var_names}

\item{response_var_used_order}{response_var_used_order}

\item{grouping_vars}{grouping_vars}

\item{grouping_var_used_order}{grouping_var_used_order}
}
\value{
power_bi_table
}
\description{
run_create_table_with_var_names_and_orders
}
