% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_text_survey_data}
\alias{select_text_survey_data}
\title{Select all of the data that should be provided in the text df}
\usage{
select_text_survey_data(data, variables_to_include)
}
\arguments{
\item{data}{data frame with the survey data}

\item{variables_to_include}{Character vector with names of the variables to keep}
}
\value{
text_survey_data
}
\description{
Select all of the data that should be provided in the text df
}
