% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_most_least_multiple}
\alias{separate_most_least_multiple}
\title{separate_most_least_multiple}
\usage{
separate_most_least_multiple(data, ml_indicator, grouping_vars)
}
\arguments{
\item{data}{data}

\item{ml_indicator}{ml_indicator}

\item{grouping_vars}{grouping_vars}
}
\value{
most_least_data_frame
}
\description{
separate_most_least_multiple
}
