% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_most_least_single}
\alias{separate_most_least_single}
\title{separate_most_least_single}
\usage{
separate_most_least_single(data, grouping_vars)
}
\arguments{
\item{data}{data}

\item{grouping_vars}{grouping_vars}
}
\value{
most_least_table
}
\description{
separate_most_least_single
}
