% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_project_working_directory}
\alias{set_project_working_directory}
\title{set_project_working_directory}
\usage{
set_project_working_directory(
  storage_platform = "dropbox",
  storage_platform_name = "TCM Dropbox",
  group_dir_name = "04 MDM Neuro-Fundraising Lab",
  jobs_folder_name = "00 Jobs",
  project_year,
  project_folder_name
)
}
\arguments{
\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}

\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}
}
\value{
working_directory_path
}
\description{
set_project_working_directory
}
