% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_project_working_directory_simple}
\alias{set_project_working_directory_simple}
\title{set_project_working_directory_simple}
\usage{
set_project_working_directory_simple(
  storage_platform_part =
    "OneDrive - Moore DM Group/Shared Documents - TCM MNFLab/00 DROPBOX IMPORT/Moore Neuro-Fundraising Lab/00 Jobs",
  project_year,
  project_folder_name
)
}
\arguments{
\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}
}
\value{
working_directory_path
}
\description{
set_project_working_directory_simple
}
