% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_up_project_environment_qualitative_coding}
\alias{set_up_project_environment_qualitative_coding}
\title{set_up_project_environment_qualitative_coding}
\usage{
set_up_project_environment_qualitative_coding(
  storage_platform,
  storage_platform_name,
  group_dir_name,
  jobs_folder_name,
  project_year,
  project_folder_name,
  should_create_nonexistant_dirs,
  survey_version_name,
  survey_monkey_used,
  wave_names
)
}
\arguments{
\item{storage_platform}{storage_platform}

\item{storage_platform_name}{storage_platform_name}

\item{group_dir_name}{group_dir_name}

\item{jobs_folder_name}{jobs_folder_name}

\item{project_year}{project_year}

\item{project_folder_name}{project_folder_name}

\item{should_create_nonexistant_dirs}{should_create_nonexistant_dirs}

\item{survey_version_name}{survey_version_name}

\item{survey_monkey_used}{survey_monkey_used}

\item{wave_names}{wave_names}
}
\value{
my_current_env
}
\description{
set_up_project_environment_qualitative_coding
}
