% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dashboard_create_fake_data.R
\name{simulate_data_from_sm_design_form}
\alias{simulate_data_from_sm_design_form}
\title{Simulate Data From the Design Form}
\usage{
simulate_data_from_sm_design_form(
  sm_design_form_path,
  desired_sample_size = 2000,
  name_of_form = "Design Form",
  name_of_logic = "Logic",
  names_of_others_to_ignore = c("IGNORE")
)
}
\arguments{
\item{sm_design_form_path}{Path to the design form}

\item{desired_sample_size}{How many responses would you like to simulate}

\item{name_of_form}{Name of the excel sheet containing the form}

\item{name_of_logic}{Name of the excel sheet containing the logic}

\item{names_of_others_to_ignore}{Name of any excel sheets that should be ignored}
}
\value{
simulated_data
}
\description{
Simulate Data From the Design Form
}
