% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum2}
\alias{sum2}
\title{Simple function to convert first to numeric and then to sum the numbers}
\usage{
sum2(x)
}
\arguments{
\item{x}{numeric variable}
}
\value{
the_sum
}
\description{
Simple function to convert first to numeric and then to sum the numbers
}
