% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_column_workbook_to_convert_vars_to_factors}
\alias{use_column_workbook_to_convert_vars_to_factors}
\title{Use column workbook to convert vars to factors}
\usage{
use_column_workbook_to_convert_vars_to_factors(
  df,
  column_details_question_type_df,
  column_workbook_list,
  question_type_indicator
)
}
\arguments{
\item{df}{data frame that contains the survey data}

\item{column_details_question_type_df}{data frame that contains the column details for a specified question type}

\item{column_workbook_list}{ist that contains the named vectors}

\item{question_type_indicator}{Character string that tells the program what data type is being evaluated and processed.}
}
\value{
df
}
\description{
Use column workbook to convert vars to factors
}
