% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wait_for_text}
\alias{wait_for_text}
\title{wait_for_text
While using R Selenium, wait until specified text appears in an element}
\usage{
wait_for_text(
  remDr,
  element,
  desired_text,
  using = "css selector",
  sleep_time = 0.1,
  give_up_count_max = 999
)
}
\arguments{
\item{remDr}{Selenium client used to control the browser.}

\item{element}{A css element as a string.}

\item{desired_text}{Specified text you are looking for}

\item{using}{which type of selector is being used}

\item{sleep_time}{second to wait between loops.}

\item{give_up_count_max}{ow many times to try the loops.}
}
\description{
wait_for_text
While using R Selenium, wait until specified text appears in an element
}
