% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_file_path_to_current_machine}
\alias{adjust_file_path_to_current_machine}
\title{adjust_file_path_to_current_machine
This code converts any path so that it works on the current computerfor the user.}
\usage{
adjust_file_path_to_current_machine(file_path)
}
\arguments{
\item{file_path}{file_path}
}
\value{
goal_file_path
}
\description{
adjust_file_path_to_current_machine
This code converts any path so that it works on the current computerfor the user.
}
