% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_file_string_with_time}
\alias{combine_file_string_with_time}
\title{combine_file_string_with_time
Function to take in a string, and then add the data and time to it}
\usage{
combine_file_string_with_time(file_string)
}
\arguments{
\item{file_string}{file_string}
}
\value{
combined_string
}
\description{
combine_file_string_with_time
Function to take in a string, and then add the data and time to it
}
