% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_specific_column_names_to_data_type}
\alias{convert_specific_column_names_to_data_type}
\title{Create a vector of data types based on column names}
\usage{
convert_specific_column_names_to_data_type(
  column_name_vector,
  text_names,
  numeric_names
)
}
\arguments{
\item{column_name_vector}{Column names vector for the dataframe}

\item{text_names}{Which column names should be labeled as text?}

\item{numeric_names}{Which column names should be labeled as numeric?}
}
\value{
data_types
}
\description{
Create a vector of data types based on column names
}
