% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_multiple_column_workbook_lists}
\alias{create_multiple_column_workbook_lists}
\title{create_multiple_column_workbook_lists}
\usage{
create_multiple_column_workbook_lists(
  my_env = my_current_env,
  column_names_paths = my_current_env$column_names_paths,
  name_of_column_details
)
}
\arguments{
\item{my_env}{my_env}

\item{column_names_paths}{column_names_paths}

\item{name_of_column_details}{name_of_column_details}
}
\value{
column_workbook_lists
}
\description{
create_multiple_column_workbook_lists
}
