% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_survey_data_for_power_bi_df}
\alias{create_survey_data_for_power_bi_df}
\title{create_survey_data_for_power_bi_df}
\usage{
create_survey_data_for_power_bi_df(
  my_env,
  write_data,
  survey_file_ext,
  survey_datetime_format_pattern,
  data_collection_survey_directory_paths,
  working_directory_path,
  column_workbook_lists,
  survey_version_name,
  convert_numeric_age_to_age_group,
  spellcheck_column_paths,
  name_of_column_details
)
}
\arguments{
\item{my_env}{my_env}

\item{write_data}{write_data}

\item{survey_file_ext}{survey_file_ext}

\item{survey_datetime_format_pattern}{survey_datetime_format_pattern}

\item{data_collection_survey_directory_paths}{data_collection_survey_directory_paths}

\item{working_directory_path}{working_directory_path}

\item{column_workbook_lists}{column_workbook_lists}

\item{survey_version_name}{survey_version_name}

\item{convert_numeric_age_to_age_group}{convert_numeric_age_to_age_group}

\item{spellcheck_column_paths}{spellcheck_column_paths}

\item{name_of_column_details}{name_of_column_details}
}
\value{
survey_data_for_power_bi_df
}
\description{
create_survey_data_for_power_bi_df
}
