% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FULL_global_coding}
\alias{FULL_global_coding}
\title{FULL_global_coding}
\usage{
FULL_global_coding(
  man_wd = NULL,
  file_part__sm_raw_folder = "Qualitative Coding/Version 2/global_variables/raw_data",
  global_coding_file_path =
    "Qualitative Coding/Version 2/global_variables/Global_Coding_USE_TO_EDIT_RESPONSES.xlsx",
  backup_for_combined_data_folder_path =
    "Qualitative Coding/Version 2/global_variables/backup_files/backup_for_combined_data",
  backup_for_editable_data_folder_path =
    "Qualitative Coding/Version 2/global_variables/backup_files/backup_for_editable_data",
  backup_for_raw_data_folder_path =
    "Qualitative Coding/Version 2/global_variables/backup_files/backup_for_raw_data",
  backup_for_data_to_add_folder_path =
    "Qualitative Coding/Version 2/global_variables/backup_files/backup_for_data_to_add"
)
}
\arguments{
\item{man_wd}{man_wd}

\item{file_part__sm_raw_folder}{file_part__sm_raw_folder}

\item{global_coding_file_path}{global_coding_file_path}

\item{backup_for_combined_data_folder_path}{backup_for_combined_data_folder_path}

\item{backup_for_editable_data_folder_path}{backup_for_editable_data_folder_path}

\item{backup_for_raw_data_folder_path}{backup_for_raw_data_folder_path}

\item{backup_for_data_to_add_folder_path}{backup_for_data_to_add_folder_path}
}
\value{
global_coding_to_save
}
\description{
FULL_global_coding
}
