% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FULL_qualitative_coding_data}
\alias{FULL_qualitative_coding_data}
\title{FULL_qualitative_coding_data}
\usage{
FULL_qualitative_coding_data(
  man_wd = NULL,
  folder_location__qualitative_coding,
  file_part__initial_name_qual_coding,
  coder_names,
  ext__initial_name_qual_coding,
  text_names,
  numeric_names,
  multiple_choice_variables,
  category_variables,
  remove_empty_sheets = TRUE,
  other_vars_that_should_not_be_counted
)
}
\arguments{
\item{man_wd}{use to add a custom working directory}
}
\value{
prepared_video_data
}
\description{
FULL_qualitative_coding_data
}
