% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_paths_and_column_data_from_excel_workbooks_list}
\alias{get_file_paths_and_column_data_from_excel_workbooks_list}
\title{get_file_paths_and_column_data_from_excel_workbooks_list}
\usage{
get_file_paths_and_column_data_from_excel_workbooks_list(
  video_coding_docs_file_paths,
  text_names,
  numeric_names
)
}
\arguments{
\item{video_coding_docs_file_paths}{video_coding_docs_file_paths}
}
\value{
file_paths_df_all_paths
}
\description{
get_file_paths_and_column_data_from_excel_workbooks_list
}
