% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initiate_credentials_for_user}
\alias{initiate_credentials_for_user}
\title{initiate_credentials_for_user
Provide Credentials for User account through the use of selenium}
\usage{
initiate_credentials_for_user(remDr)
}
\arguments{
\item{remDr}{Selenium client used to control the browser}
}
\value{
remDr
}
\description{
initiate_credentials_for_user
Provide Credentials for User account through the use of selenium
}
