% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_column_name_paths_for_all_survey_versions}
\alias{obtain_column_name_paths_for_all_survey_versions}
\title{obtain_column_name_paths_for_all_survey_versions}
\usage{
obtain_column_name_paths_for_all_survey_versions(version_directories)
}
\arguments{
\item{version_directories}{version_directories}
}
\value{
version_column_names_paths
}
\description{
obtain_column_name_paths_for_all_survey_versions
}
