% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_newest_file_per_directory}
\alias{obtain_newest_file_per_directory}
\title{For every directory, obtain the newest file}
\usage{
obtain_newest_file_per_directory(
  version_directories,
  file_ext = ".zip",
  format_pattern = "_[0-9]{8}_[0-9]{4}"
)
}
\arguments{
\item{version_directories}{version_directories}

\item{file_ext}{file_ext}

\item{format_pattern}{format_pattern}
}
\value{
newest_files_for_project
}
\description{
For every directory, obtain the newest file
}
