% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_data_for_calcs}
\alias{prepare_data_for_calcs}
\title{prepare_data_for_calcs}
\usage{
prepare_data_for_calcs(sheet_coding_data)
}
\arguments{
\item{sheet_coding_data}{sheet_coding_data created with `process_video_data`}
}
\value{
wide_data_for_calcs
}
\description{
prepare_data_for_calcs
}
