% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Content_Evaluation_Qualitative_Coding.R
\name{process_video_data}
\alias{process_video_data}
\title{process_video_data}
\usage{
process_video_data(file_paths_df_all_paths, column_workbook_list)
}
\arguments{
\item{file_paths_df_all_paths}{file_paths_df_all_paths}

\item{column_workbook_list}{column_workbook_list}
}
\value{
sheet_coding_data_df
}
\description{
process_video_data
}
