% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_video_data}
\alias{process_video_data}
\title{process_video_data}
\usage{
process_video_data(
  video_coding_docs_file_paths,
  text_names,
  numeric_names,
  multiple_choice_variables,
  category_variables
)
}
\arguments{
\item{video_coding_docs_file_paths}{video_coding_docs_file_paths}
}
\value{
sheet_coding_data_df
}
\description{
process_video_data
}
